+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file. This change
 
 ## [Unreleased]
 
+## 0.4.0 (2019-12-11)
+
 - [BREAKING] Schema :[function,multimethod,spec.alpha,fspec.alpha]/:source-form collapsed into :source/form, same with :source/str
 - Fix multimethod methods facts to correctly link methods with its multi definition